requirements.txt
streamlit
pandas
plotly
python-dotenv
openai
fpdf
numpy
chromadb